12_base_tool/01_market_scanner/requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# API clients
polygon-api-client>=1.12.0

# Excel integration
xlwings>=0.30.0

# Configuration
python-dotenv>=1.0.0

# Market timing
pytz>=2023.3
pandas-market-calendars>=4.3.0

# Web scraping (for ticker updates)
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database (optional - for Supabase)
supabase>=2.0.0

# Logging and utilities
typing-extensions>=4.5.0

# Database
aiohttp>=3.9.0
pyarrow